fix: get own post api
